Bug 1056516 - (wip) let auto hyphen honor manual hyphen when hyphens:auto is set.
Done:
1. use StyleHyphens enum class to avoid using multiple arrays
2. handle the explicit hyphen cases
3. add tests for the explicit hyphen cases

Todo:
1. replace isWordBoundary with a better approach
2. handle the case of the text being processed exceeds the MEASUREMENT_BUFFER_SIZE

According to CSS Text 3 - 6.1. Hyphenation Control:
Automatic hyphenation opportunities within a word must be ignored if the word
contains a conditional hyphen. After breaking at such opportunities, if a
portion of that word is still too long to fit on one line, an automatic
hyphenation opportunity may be used.

We can't get the info about if there exists soft hyphen(s) in a word unless we
get to PropertyProvider::GetHyphenationBreaks. At that timing, we also import
the auto hyphen info, which means we can get both soft hyphen and auto hyphen
info there. That's the reason why I'm putting the detection logic in the same
place. Once we get the extra info that we need, we can do the honoring thing in
BreakAndMeasureText accordingly.

MozReview-Commit-ID: 60T02Z58NRs