Bug 1348173 - update test expectations for -moz-border-*-colors support.
With the support of -moz-border-*-colors (see previous parts of the patchset),
165 mochitests and 17 reftests could be fixed.

This is just for the convenience of review. Once this patch get r+, I'll merge
this part with other Gecko changes into one single commit, so we could minimize
the risk of breaking autoland repo while doing the Servo/Gecko syncing stuff.

MozReview-Commit-ID: INjUW0tbDnt