Bug 1359313 - whitelist -moz-border-*-colors related bindings to satisfy heap write analysis.
A quick fix for hazard bustage by increase the NUM_ALLOWED_WRITE_HAZARDS
from 3 to 7 is pushed in bug 1348173 comment 37. In this bug, we shall do
the actual fix and restore the NUM_ALLOWED_WRITE_HAZARDS.

The -moz-border-*-colors bindings trigger errors because they're using
outparams (nsStyleBorder) which further manipulate its member (mBorderColors)
which is a double raw pointers. We might remove these from the whitelist
if we could refactor mBorderColors to use nsTArray directly.

MozReview-Commit-ID: 2cQz58K2A10